refactor logic out of main.go
