+++ deputy-claimer/Makefile
@@ -1,0 +1,6 @@
+install:
+	go install
+
+test_integration:
+	cd integration_test && run-tests.sh
+
