+++ deputy-claimer/integration_test/run-tests.sh
@@ -1,27 +1,29 @@
-# #! /bin/bash
+#! /bin/bash
 
-# # Note: this script should be run from within its directory (otherwise it can't find the docker-compose.yaml file)
+# Note: this script should be run from within its directory (otherwise it can't find the docker-compose.yaml file)
 
-# # exit on errors
-# set -e
+# exit on errors
+set -e
 
-# # start up the deputy and chains
-# docker-compose up -d
+# start up the deputy and chains
+docker-compose up -d
 
-# # wait until the deputy is operational
-# echo "waiting for deputy to start"
-# while ! docker-compose exec deputy curl --fail localhost:8080/status > /dev/null
-# do
-#     sleep 1
-# done
-# echo "done"
+# wait until the chains are operational
+echo "waiting for kava node to start"
+while ! docker-compose exec kavanode curl --fail localhost:26657/status > /dev/null
+do
+    sleep 1
+done
+# wait until a block is committed
+sleep 4
+echo "done"
 
-# # run tests
-# # don't exit on error, just capture exit code (https://stackoverflow.com/questions/11231937/bash-ignoring-error-for-a-particular-command)
-# # use -count=1 to disable test result caching
-# go test . -count=1 -tags integration -v && exitStatus=$? || exitStatus=$?
+# run tests
+# don't exit on error, just capture exit code (https://stackoverflow.com/questions/11231937/bash-ignoring-error-for-a-particular-command)
+# use -count=1 to disable test result caching
+go test . -count=1 -tags integration -v && exitStatus=$? || exitStatus=$?
 
-# # remove the deputy and chains
-# docker-compose down
+# remove the deputy and chains
+docker-compose down
 
-# exit $exitStatus
+exit $exitStatus
